% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReSEM.R
\name{assign.ReSEM.sampling}
\alias{assign.ReSEM.sampling}
\title{Rejective Sampling}
\usage{
assign.ReSEM.sampling(num.assign, N, n, W, prob.S)
}
\arguments{
\item{num.assign}{Number of assignments to be generated.}

\item{N}{Size of the finite population.}

\item{n}{Number of the sampled units.}

\item{W}{An \code{N} by J matrix which contains J dimensional covariate vector for each unit, available at the sampling stage.}

\item{prob.S}{Asymptotic acceptance probability for the sampling stage.}
}
\value{
A list with the elements
\item{ind.assign}{A \code{num.assign} by \code{N} sampling matrix. Each row of the sampling matrix is an \code{N}-dimensional vector that represents one sampling result, where we use 0 for sampled units, and -1 for units that will not enter the experiment. }
\item{M.S.all}{An vector of length \code{num.assign}, which stores the Mahalanobis distances corresponding to each sampling. }
}
\description{
Generates sampling indicators under rejective sampling using the Mahalanobis distance.
}
