% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReSEM.R
\name{assign.ReSEM.treatment}
\alias{assign.ReSEM.treatment}
\title{Rerandomized treatment assignment}
\usage{
assign.ReSEM.treatment(n, n1, X, prob.T)
}
\arguments{
\item{n}{Number of units to assign.}

\item{n1}{Number of the treated units.}

\item{X}{An \code{n} by K matrix which contains K dimensional covariate vector for each unit, available at the treatment assignment stage.}

\item{prob.T}{Asymptotic acceptance probability for the treatment assignment stage.}
}
\value{
A list with the elements
\item{ind.assign}{An assignment vector of length \code{n}, where we use 1 for treated units, and 0 for control units. }
\item{M.T}{The corresponding Mahalanobis distance. }
}
\description{
Generates treatment-control assignment under rerandomization using the Mahalanobis distance.
}
