% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReSEM.R
\name{dif.reg}
\alias{dif.reg}
\title{Regression-adjusted difference-in-means}
\usage{
dif.reg(assignment, Y, C, E)
}
\arguments{
\item{assignment}{Treatment assignment vector, which contains -1, 0, 1 values. We use 1 for treated units, 0 for control units, and -1 for units that will not enter the experiment.}

\item{Y}{Observed outcomes for all units. Will be NA for units that will not enter the experiment.}

\item{C}{Available covariate vector for sampled units, at the analysis stage.}

\item{E}{Available covariate vector for all units, at the analysis stage.}
}
\value{
Regression-adjusted difference-in-means estimator for the average treatment effect, with estimated adjustment coefficients.
}
\description{
Calculates the regression-adjusted difference-in-means estimator for the average treatment effect.
}
