% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReSEM.R
\name{diff.in.means}
\alias{diff.in.means}
\title{Difference in means}
\usage{
\method{diff}{in.means}(assignment, Y)
}
\arguments{
\item{assignment}{Treatment assignment vector, which contains -1, 0, 1 values. We use 1 for treated units, 0 for control units, and -1 for units that will not enter the experiment.}

\item{Y}{Observed outcomes for all units. Will be NA for units that will not enter the experiment.}
}
\value{
Difference-in-means estimator for the average treatment effect.
}
\description{
Calculates the difference-in-means estimator for the average treatment effect.
}
