% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReSEM.R
\name{generate.constrained.Gaussian}
\alias{generate.constrained.Gaussian}
\title{Generate constrained Gaussian}
\usage{
generate.constrained.Gaussian(num, K, a)
}
\arguments{
\item{num}{Number of constrained Gaussian random variables to generate.}

\item{K}{Dimension of the Gaussian vector used to generate constrained Gaussian random variable.}

\item{a}{Threshold used to generate constrained Gaussian random variable.}
}
\value{
A vector of length \code{num}, which contains iid constrained Gaussian random variables.
}
\description{
Generates constrained Gaussian random variables.
}
