% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReSEM.R
\name{obs.outcome}
\alias{obs.outcome}
\title{Observed outcomes}
\usage{
obs.outcome(assignment, Y1, Y0)
}
\arguments{
\item{assignment}{Treatment assignment vector, which contains -1, 0, 1 values. We use 1 for treated units, 0 for control units, and -1 for units that will not enter the experiment.}

\item{Y1}{Potential outcomes under treatment.}

\item{Y0}{Potential outcomes under control.}
}
\value{
A vector of the same length as \code{assignment}, \code{Y1} and \code{Y0}, which contains observed outcomes for all units. If a unit will not enter the experiment, its observed outcome will be NA.
}
\description{
Computes observed outcomes using potential outcomes and treatment assignment.
}
